Minor test code cleanup
